Add an option to show all failed tests
As per n3337 3.5.3,
a name having namespace scope has internal linkage if it is the name of

- a variable, function or function template that is explicitly declared
`static`; or,

- a variable that is explicitly declared `const` or `constexpr` and
neither explicitly declared `extern` nor previously declared to have
external linkage; or

- a data member of an anonymous union.

So 'show_all_failed_tests' must be declared `extern` in its definition.

Note that 'show_all_failed_tests' is declared `extern constexpr`
in task.cpp to ensure that there is a definition.