Simplify implementation of vv<T>
Let vv<T> inherit std::vector<std::vector<T>>.
vv<T>'s member 'n' is default-initialized to indeterminate value when
vv<T> is default initialized.

Something new that I learned about C++:

Base class is initialized first, regardless of the order it is mentioned
in the constructor initializer list.

Since C++03, default initialization is performed in three situations:
1) when a variable with automatic, static, or thread-local storage
duration is declared with no initializer; (in the form 'T object;')
2) when an object with dynamic storage duration is created by
new-expression with no initializer; (in the form 'new T')
3) when a base class or a non-static data member is not mentioned in a
constructor initializer list and that constructor is called.

Since C++11, the effects of default initialization are:
- if T is a class type, the constructors are considered and subject to
overload resolution against the empty argument list. The constructor
selected (which is one of the default constructors) is called to provide
the initial value for the new object;
- if T is an array type, every element of the array is
default-initialized;
- otherwise, nothing is done: the objects with automatic storage
duration (and their subobjects) are initialized to indeterminate values.

Reference:
https://en.cppreference.com/w/cpp/language/default_initialization