Optimize the naive algorithm for max bi-matching
This algorithm searches for ONE augmenting path from each vertex on the
left part. Naive implementations of this algorithm are slow in some
cases.

I found out that sorting nodes on the left part by degree is shuffling
those nodes in effect. Without a shuffle, the algorithm will TLE on
chain because each time it will first find a long non-augmenting
alternating path before it finds an augmenting path of length 1, taking
an overall O(N^2) time.

The idea of my optimization is to exclude vertices that we know will
never lead to an uncovered vertex on the right part. It passed all tests
in Library-Checker, see https://judge.yosupo.jp/submission/26518. Note
that the code is not well tested and may be wrong and can be further
optimized.

There is a variant of the augmenting path algorithm which finds multiple
(possibly) short augmenting paths on each phase using DFS. It is easy to
implement and is fast. See https://judge.yosupo.jp/submission/22648.