Use 256 MiB stack
Quote from https://sourceware.org/binutils/docs/ld/Options.html#Options
--stack reserve
--stack reserve,commit
Specify the number of bytes of memory to reserve (and optionally commit) to be used as stack for this program. The default is 2MB reserved, 4K committed. [This option is specific to the i386 PE targeted port of the linker]

I am using MSYS2's distribution of gcc, and the default target of ld is
i386pe. You can see this in the output of `gcc -dumpspecs`.